move DefaultCameraErrorListener to camera-core
