+++ camera-core/res/values/strings.xml
@@ -3,5 +3,10 @@
     <string name="stripe_camera_permission_denied_message" description="Message shown to the user when camera permission is denied">Please allow camera access to scan your card</string>
     <string name="stripe_camera_permission_denied_ok" description="Affirmative button shown as part of the permission denied dialog">OK</string>
     <string name="stripe_camera_permission_denied_cancel" description="Negative button shown as part of the permission denied dialog">Cancel</string>
-    
+
+    <string name="stripe_error_camera_title" description="Title of the dialog shown to the user when an error using the camera occurs">Camera Problem</string>
+    <string name="stripe_error_camera_open" description="Message of the dialog shown to the user when an error starting the camera occurs">The camera failed to turn on</string>
+    <string name="stripe_error_camera_access" description="Message of the dialog shown to the user when an error accessing the camera occurs">Permission was denied when turning on the camera</string>
+    <string name="stripe_error_camera_unsupported" description="Message of the dialog shown to the user when the camera is not supported">This device does not support the required camera features</string>
+    <string name="stripe_error_camera_acknowledge_button" description="Affirmative button shown as part of the camera error dialog">Close</string>
 </resources>
